Implement parser for groff mom files
This is a necessity because we want to support producing manuscripts in
the docx format. Parsed manuscripts are now represented as a slice of
tokens, which will either be of type Macro, delineating a groff macro
such as ".PP", or of type Text, that is any text that is not a Macro.

With these tokens in place we essentially parse them when publishing to
the docx format and use this to determine the markup for the final
document. The final document will not have parity with the pdf format,
but it will be good enough for submission. That is, Times New Roman,
font size 12pt, and double spaced.

For producing pdf files we still use "groff -T pdf -k -mom" under the
hood.