Handle both CHAPTER and CHAPTER_TITLE macros
CHAPTER is used to set the chapter number, this can either be a digit, a
numeral, or a word. CHAPTER_TITLE is used to set the chapter title and
will be displayed beneath the number of both CHAPTER and CHAPTER_TITLE
are given within a manuscript.

Either can be used to delineate the start of a chapter, whilst they can
be used interchangeably this will lead to inconsisten manuscript
rendering for both PDF and DOCX formats.