Make copies of the tokens for each chapter we parse
This simply makes things easier to work with, and a little bit of
duplication in memory isn't going to be too costly. This also makes
publishing individual chapters more reliable too, since we aren't
keeping track of the token positions any more, and have the actual
tokens themselves to rely on.